refactor(playwright): separate TestForm, co-locate script types, remove unnecessary exports, wrap tests in panes, add placeholder array input spec
