+++ packages/sanity/playwright-ct/tests/utils/testHelpers.tsx
@@ -93,54 +93,3 @@ export function testHelpers({
     },
   }
 }
-
-/**
- * Maps a test status in Browserstack to a session status.
- * @param status - The test status.
- * @returns The corresponding session status ('passed' or 'failed').
- */
-function mapTestStatusToSessionStatus(status: string) {
-  const successStatuses = ['passed', 'skipped'] //
-  return successStatuses.includes(status) ? 'passed' : 'failed'
-}
-
-/**
- * Sets the session status in Browserstack based on the test status and error information.
- * @param page - The Playwright test page.
- * @param testInfo - Information about the current test.
- */
-export async function setSessionStatus(page: PlaywrightTestArgs['page'], testInfo: TestInfo) {
-  const {status, error} = testInfo
-  const sessionStatus = mapTestStatusToSessionStatus(status || 'failed')
-
-  // eslint-disable-next-line no-nested-ternary
-  const reason = error ? error.message : status === 'skipped' ? 'skipped' : 'n/a'
-
-  const executionCommand = `browserstack_executor: ${JSON.stringify({
-    action: 'setSessionStatus',
-    arguments: {
-      status: sessionStatus,
-      reason: reason,
-    },
-  })}`
-
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  await page.evaluate((_) => {}, executionCommand)
-}
-
-/**
- * Sets the session test name in Browserstack based on the test information.
- * @param page - The Playwright test page.
- * @param testInfo - Information about the current test.
- */
-export async function setSessionTestName(page: PlaywrightTestArgs['page'], testInfo: TestInfo) {
-  const executionCommand = `browserstack_executor: ${JSON.stringify({
-    action: 'setSessionName',
-    arguments: {
-      name: testInfo.title,
-    },
-  })}`
-
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  await page.evaluate((_) => {}, executionCommand)
-}
